API Sidebar fixed to link to correct IDs
The Rails helper function generates an ID by parameterizing the header's
text. This breaks if there are numerous headers the same (real life
examples: "Request", "Response", "Keys and Values" etc.)

Instead, we can just read the ID that's been generated by the Markdown
parser and use that instead. This should fix lots of broken links on the
site.