Tabbed Examples Filter now throws an error if source/tabs not set
The point of this is to make it so an obvious error is thrown when the
filter does not get the information it needs to render the template.

The preconditions of the filter is that the YAML block contains either
tabs or source. If it doesn't contain either, we should throw an error
earlier rather than wait for a bigger problem to bubble up.