Render all documentation pages at build time in case exceptions are thrown
In a recent commit a change was made that added a new `type` field to
building blocks without a default. The decision to fail fast was made as
it looked like all building blocks had been updated.

It wasn't until we tried to update the search index (which renders every
page) that we noticed an exception being thrown by a building block that
hadn't been updated.

This commit adds a build step that renders every markdown file at build
time to prevent this happening again