Use correct terminology for `app` connect in NCCO (#1602)
We should not use the term 'ip leg' externally. 

The leg type: app (both in the code and in the docs). no need to mention 'legs' here. we connect TO users VIA legs(legs is the means). we don't connect to legs :)
can either use: application or app

I used app here so that people don't get confused with the concept of Nexmo application :/

We connect to a denoted user, not a member. the member is created as a result of this ncco :) the vapi user is also unaware of the concept of the member so lets not add confusion ;)