Add ci:verify_error_urls_resolve rake task
Adds a CI check to make sure that all error messages used in the Open API documents have a corresponding entry on Nexmo Developer.

This PR will fail CI until additional errors are added to OAS docs (I'll resolve those issues before this PR is merged).

This recently happened when we merged the account documentation to a single OAS. The production API type field can not change, so we added additional routing to make sure that the URLs still resolved correctly.