[DEVX-3608] Set the same value for <meta description> and <og:description>.
If there is a `@frontmatter` present these tags will use the values of
the following keys:
  * 'meta_description'
  * 'description'
in that specific order.

In case the `@frontmatter` is not defined, it will use the value of the
`meta_description` key defined in `config/header_meta.yml`. This way,
the landing page will use the correct values, and docs will use whatever
is defined in their frontmatter having the config file as a fallback.